requirements.txt
future~=0.18.3
customtkinter~=5.2.2
pillow~=10.2.0
pandas~=1.5.3
matplotlib~=3.7.0
numpy~=1.26.4
joblib~=1.4.0
plotly~=5.21.0
Pmw~=2.1.1